evaluators/requirements.txt
langchain-openai>=0.1.0
pandas>=1.5.0
numpy>=1.24.0
jinja2>=3.1.0
nltk>=3.8.0
rouge-score>=0.1.0
promptflow-core>=1.0.0
python-dotenv>=1.0.0
langchain-core>=0.1.0
langchain-ollama>=0.1.0 